requirements.txt
gradio>=4.0.0
gradio-client>=0.10.0
pandas>=2.0.0
google-generativeai>=0.3.0
pypdf>=3.0.0
loguru>=0.7.0
promptic>=0.1.0
tenacity>=8.2.0
requests>=2.31.0
gTTS>=2.3.0
pyttsx3>=2.90
edge-tts>=6.1.0